+++ tests/Jroc.Tests/Hosting/JavaScript/Hosting_RootFunctionExport.js
@@ -1,0 +1,8 @@
+"use strict";
+
+function rootIncrement(value) {
+    return value + 1;
+}
+
+rootIncrement.kind = "root-function";
+module.exports = rootIncrement;